refactor: DTO 패키지 조정 및 네이밍 변경
