+++ src/main/java/wooteco/subway/admin/controller/StationController.java
@@ -3,8 +3,8 @@ package wooteco.subway.admin.controller;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import wooteco.subway.admin.domain.Station;
-import wooteco.subway.admin.dto.StationCreateRequest;
-import wooteco.subway.admin.dto.StationResponse;
+import wooteco.subway.admin.dto.request.StationCreateRequest;
+import wooteco.subway.admin.dto.response.StationResponse;
 import wooteco.subway.admin.repository.StationRepository;
 
 import java.net.URI;
