+++ src/main/java/wooteco/subway/admin/dto/request/LineCreateRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.admin.dto;
+package wooteco.subway.admin.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import wooteco.subway.admin.domain.Line;
@@ -6,7 +6,7 @@ import wooteco.subway.admin.domain.Line;
 import javax.validation.constraints.Pattern;
 import java.time.LocalTime;
 
-public class LineRequest {
+public class LineCreateRequest {
     private String name;
     @Pattern(regexp = "(?:[01]\\d|2[0123]):(?:[012345]\\d):(?:[012345]\\d)\n", message = "hh:mm:ss 의 형태가 아닙니다.")
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "HH:mm:ss")
@@ -17,7 +17,7 @@ public class LineRequest {
     private int intervalTime;
     private String lineColor;
 
-    public LineRequest() {
+    public LineCreateRequest() {
     }
 
     public String getName() {
